LayoutTests/fast/regions/cssom/get-region-flow-ranges-writing-modes-tb-rl-lr-expected.txt

[CSSRegions] Test for Region::getRegionFlowRanges for mixed vertical modes
PASS range1.length is 2
PASS Array [wrapper,0,box1,1] is equal to [wrapper,0,box1,1]
PASS Array [box3,1,box3,2] is equal to [box3,1,box3,2]
PASS range2.length is 1
PASS Array [box1,1,box1,2] is equal to [box1,1,box1,2]
PASS range3.length is 1
PASS Array [box1,2,box3,1] is equal to [box1,2,box3,1]
PASS successfullyParsed is true

TEST COMPLETE
Vertical RL
X
Vertical LRY
